requirements.txt
pdfplumber
pandas
PyQt6
numpy
scikit-learn
sentence-transformers
python-dotenv
PyPDF2
openpyxl
requests